+++ packages/create-app/templates/react-aigne-dapp/api/dev.ts
@@ -1,0 +1,7 @@
+import { app, server } from './src';
+
+import('vite-plugin-blocklet').then(({ setupClient }) => {
+  setupClient(app, {
+    server,
+  });
+});