+++ packages/create-app/templates/react-aigne-dapp/api/third.d.ts
@@ -1,0 +1,17 @@
+declare module 'vite-plugin-blocklet';
+
+declare module 'express-history-api-fallback';
+
+declare module 'express-async-errors';
+
+namespace Express {
+  interface Request {
+    user?: {
+      did: string;
+      role: string;
+      fullName: string;
+      provider: string;
+      walletOS: string;
+    };
+  }
+}