+++ packages/create-app/templates/react-aigne-dapp/src/contexts/session.tsx
@@ -1,0 +1,11 @@
+import { createAuthServiceSessionContext } from '@arcblock/did-connect/lib/Session';
+import { useContext } from 'react';
+
+const { SessionProvider, SessionContext, SessionConsumer, withSession } = createAuthServiceSessionContext();
+
+function useSessionContext() {
+  const info = useContext(SessionContext);
+  return info as any;
+}
+
+export { SessionProvider, SessionContext, SessionConsumer, useSessionContext, withSession };